Get involved: Allow scrollbars and smaller overlay
Note: That doesn't work very good on an 1024x768 screen because
of the size of the content. Maybe we should opt for a different
markup here, or tweak it to not use elements with id's that re
already used on the page (#col1, #col3) and shrink the width
so the overlay and it's content fit's into 1000px width.