first hierarchical norms listing implementation
This contains a hack as it relies to the number of the sort to adjust the
pages result set to the top level pages, and pass information about the
rendering to the tile via the tmpl_context variable. Downside: The number
of the sort changes across languages, so the listing doesn't work correctly
in other languages. This can only be fixed in the pager.