Proposal adoption: Do not adopot the variants title
We do not allow to edit the title of a variant (not the name of a variant)
but leave it untouched. If we edit a norm and change the title, variants created
earlier keep the old title (hidden to the user). If we adopt the title of
the variant we would set back the changes.