badge controller hack: manual validation without @validator
Fixes Errors if a form failed to validate with exceptions like::

     (DataError) invalid input syntax for integer: "test"

The @validate decorator dispatches to the given form. This expects
that the form accepts the same arguments as the update method that
was called by routes with a matching route.

In our case we have two different routes, e.g::

  /badge/edit/{id}
  /instance/{id}/settings/badges/edit/{badge_id}

and we manually dispatch from the latter action to the action of the
former. As you can see the signatures of both (and their corresponding
controller actions) are different. {id} is in the former case the
badge id, and in the latter the instance id. But the @validate
decorator pathes the arguments figured out by route to the form
method.

The latter route /instance/... in the InstanceController will get passed
two arguments, the instances id and the badge_id, e.g.::

  *('test', '37')

and the form action in the BadgeController has a different signature::

  BadgeController.add(self, id, errors=None)

cause it's not in an instance. But will be called by @validator with::

  BadgeController.add('test', '37')

instead of what he expects::

  BadgeController.add('37', ...)

So it treats 'test' like a badge id.

This commit makes sure the form methods are called with the right
arguments.