Allow to increase maximal instance key size from 20 to 63 characters
This lifts the 20 characters maximum instance key restriction on
database level to 63 characters; this is the maximum length of a label
of a node of a FQDN (meaning any part between dots in the domain name),
according to RFC 1034 [0].

The default setting of the configuration option
adhocracy.instance_key_length_max is still 20 characters, but can now be
increased up to 63.

Some examples:

* The longest currently used instance key has 35 characters:
stadtbremisches-ueberseehafengebiet (geo branch has allowed this for a
while).

* The longest German municipality title "Vereinbarte
Verwaltungsgemeinschaft der Stadt Geislingen an der Steige" wouldn't be
covered by the 63 characters limit, as it has a length of 70 characters.

[0] http://tools.ietf.org/html/rfc1034#section-3.1