Feedback creation UI (fixes #88)
Second part of #88: If configured, permanently show a feedback button on
the right side, which on click makse a modal form slide in, which allows
to enter feedback, which ends up as in a new proposal in the feedback
instance.

With this commit, the form uses the normal proposal creation form of the
feedback instance as submit target. On validation or successful submit,
the user is thus redirected to the feedback instance.

In a next step, form submission and validation should be done through
AJAX.