Allow to span multiple sections in "show more"
Use python-markdown preprocessor instead of inlinepattern, as
inlinepattern only deal with single blocks. It is now possible to hide
multiple sections in a "show more" block.

The [More...] link is now pushed to next line, as the "show more" magic
is now done with `<div>` elements instead of `<span>`, which makes
markdown close the previous `<p>` element, as `<p>` only allows inline
elements inside.

In order to resolve this, some more brainpower has to be used...