Upgrade Pylons (0.9.7 -> 1.0.1)
Adjust code to the upgrade, as described in
http://docs.pylonsproject.org/projects/pylons-
webframework/en/latest/upgrading.html

Due to an apparent incompatibility between WebOb>=1.1.1 and Paste-
1.7.5.1, or some oddness in the Adhocracy setup code, a workaround in
`i18n/__init__.py` has to be introduced in order to catch an exception.

While at it, some depending libraries and other non-depending libraries
are updated on the fly:
 * WebOb (1.0.8 -> 1.2.3)
 * repoze.who-friendlyform (1.0.4 -> 1.0.8)
 * IPython (0.11 -> 0.13.1)
 * ipdb (0.5 -> 0.7)
 * Routes (1.12.3 -> 1.13)
 * PasteScript (1.7.3 -> 1.7.4.2)
 * MarkupSafe (0.12 -> 0.15)
 * Beaker (1.5.4 -> 1.6.4)

See changed version file of corresponding commit in adhocracy.buildout.

This now also allows to upgrade other dependencies, such as fanstatic.

This upgrade also removes the need for the `paster shell` command patch
in order to get a proper IPython shell as described in #74.