Allow assignment of categories to pages
It is now possible to assign categories to pages. The categories to be
assigned are the same categories as the proposal categories, as these
CategoryBadges are DelegateableBadges anyway, and can thus be used for
pages as well.

This way these categories may act as another bridge between proposals
and pages. If this isn't always desired, additional attributes can be
added later to the CategoryBadge model in order to define, whether a
category is meant as a proposal category, a badge category or both.

This commit doesn't add category facets to the page listing.