Add user rating algorithm and sorting (fixes #190)
An algorithm is created which sums ratings user receive for their
contributions. This currently simply adds up the score for user's
comments and the score for the user's proposals rate polls.

This score is added to the user documents in the solr index as separate
fields per instance, as it is done with user activity.

Consequently, the solr score can be used for rating based sorting of the
instance membership lists.