Only trust defined OpenID providers
As everybody can install an OpenID provider, this commit adds a list of
trusted OpenID providers.

This is currently Google, Yahoo and MyOpenID, but can easily be
extended.

Only email addresses of trusted OpenID providers are automatically
validated, while others will receive an email with a confirmation link
as after normal registrations.

Note that it is still possible to create an account without email
address, if the OpenID provider returns none.