Match OpenID emails to existing users
If an email address provided by OpenID attribute exchange upon on-
the-fly OpenID signup is already used by another user:

1. Do not simply create a new user with the same (duplicate) email
address (as written in a comment in #172), but

2. Automatically log in the user with that email address assigned, if
the OpenID provider is trusted, and

3. Do not log in or sign up, but issue a warning, if this happens with a
non trusted OpenID provider. Otherwise this would allow to login as any
Adhocracy user by faking an OpenID provider.