Add another jquery quirk to fix dynamic overlay
In some cases, getTrigger() returns a jquery object, which returns a
title when using `.data('title')` (e.g. add comment), while in other
cases, `.attr('title')` must be used (e.g. propose new norm).

As I don't find the reason for this (this might perfectly be a bug in
jquery / jquery-tools), we just check for both.