Fix fake_job method name
FakeJob is supposed to work in cases where redis isn't available.

Unfortunately, this doesn't work, as the `Job` constructor and other
methods has changed in newer versions of `rq`.

Fallback to `sync` if redis isn't available would be nice-to-have, but
isn't critical for the RabbitMQ replacement, as the availability of
RabbitMQ is currently mandatory as well.