Allow formatted norms
This commit allows pages to be explicitly marked as formatted.

This currently only has an effect on pages, which are norms ; pages,
which are proposal descriptions, are still always rendered with
markdown. Furthermore, formatting cannot be deactivated on proposal
descriptions through the normal UI workflow.

If formatting is enabled, norms rendered with markdown in the page view
instead of as plain text in a line-by-line view.

If diff is activated, the line-by-line view is used and the changes are
applied to the text as if formatting was off. At a later stage, the
markdown can either be stripped away from the diff, or a real html-diff
can be used.