Add option to monitor how much time a user spends on a page
An option 'adhocracy.monitor_page_time_interval' is introduced. If
set to a positive number, background requests are send to the stats
controller to monitor how long a page stays opened in the browser.
It is also included whether the window/tab is currently focused.
The value of the property determines the time between the completion of
the previous background request and the start of the next one.

This feature is required for hhucn/adhocracy.hhu_theme#167.