Makek massmessage footer optional.
Currently, every message has the unsubscription footer. That's fine (even required) in open communities, but out of place if there is some kind of central user management (say, in a company), and if non-technical users may not even be aware we're using adhocracy as a newsletter service.
This commit therefore allows users with global.admin permissions to suppress the footer.