Fix translation when en is first element (Fixes hhucn/adhocracy.hhu_theme#271)
This fixes a misuse of a Pylons-specific method that is itself broken: request.languages should just return an ordered list of requested languages, but actually replaces everything that looks like the default language (en-us); *and all the languages thereafter*; with the (invalid) default language, which means that a language preference of en,de will evaluate to de. Also, the languages property will never return an empty value; which is not what we want (we fall back to the installation-specific default languages) and only worked on accident because of the aforementioned bug.