Allow custom sender name in massmessages (Fixes hhucn/adhocracy.hhu_theme#256)
The initial impression of an email may not only depend on the sender address, but also the name.
We should therefore allow users with global.message privileges to customize the sender name.
The functionality is already there, we just need an input field and a database row to store the sender name in.