Store session values in cookies.
Currently, we (or more precisely, beaker)  store session values in the filesystem.
That does not scale, and is extremely hard to synchronize across multiple machines.
Therefore, this code adds a new flag adhocracy.session.implementation that can be set to "cookie" to use a new session implementation which stores the session data in cookies.
The current implementation does not encrypt the stored data (as I haven't found any secret values), but does check a signature in order to prevent users from meddling with these values.