Cleaner permission initialisation
Previously, permissions were only assigned to the corresponding default
groups on initial installation, but not if they are introduced later.

This commit introduces the configuration option
adhocracy.autoassign_permissions, which automatically assigns newly
introduced permissions to the default groups. This option is enabled by
default and should be turned off in restrictive environments.