Reduce Shibboleth column length to 255 characters
This is required because MySQL doesn't support indexes on tables larger than 767 bytes.
(See http://dev.mysql.com/doc/refman/5.5/en/innodb-restrictions.html for documentation on that).

The current limit of 1024 was arbitrary anyways - according to http://www.switch.ch/aai/support/serviceproviders/persistentid.html , the real limit is 2036 characters.
This change fixes #390.
What we should really do - and this seems to be pondered by Shibboleth protocol designers anyways - is to hash the persistent ID, and to index over that.
If needed, we can still store the original long ID in the database.