Log the list index and active sort of pager clicks
When a user clicks a proposal the list index and active sort are logged
into a cookie. On the next page load, this cookie is read and the data
is send to the stats controller. The option
'adhocracy.monitor_pager_clicks' must be set to True to enable the
functionality.

This is needed for hhucn/adhocracy.hhu_theme#119.