Add badge-specific behaviors
If the new configuration option adhocracy.enable_behaviors is enabled, user badges can be associated with changes in the UI.
The first UI change is the default sorting order.
Instead of defaulting to mixed, if the user has a badge with a sorting order behavior attached, she now gets that order.

See hhucn/adhocracy.hhu_theme#121 for details.