Request adhocracy_ko resource outside of tile (fixes #427)
Tile caching can leed to fanstatic resources not being requested if a
template isn't rendered but taken from the cache.

This happens in the show proposal page, which includes selection tiles,
which need adhocracy_ko, which is missing if the tiles are taken from
the cache.

This commit therefore makes sure the fanstatic resource isn't requested
from within the tile.