Fix changing norms from within proposals (fixes #406)
The behavior of PickleType has been changed in SQLAlchemy 0.7 such that
changes of such attributes aren't detected automatically anymore. This
commit restores the previous behavior.

This needs to be touched again when moving to SQLAlchemy 0.8, as the
mutable flag will then be removed and we have to create a custom data
type for the variants list, which inherits from the `Mutable` mixin.