Track outgoing links
Currently, we do not register when a user clicks on a link to another page (say, information about the project, or a link to additional data / similar projects).
Similar to how google and facebook do it, this change makes all links in static page go to an intermediary redirect site, and then get redirected to the final URL.
Fixes hhucn/adhocracy.hhu_theme#341.