Add a configuration to not check for the feedback instance on every request
Currently, we're sending a database query for every single rendering of adhocracy, just for the feedback button, and just to verify that the instance exists.
Add a configuration option adhocracy.feedback_check_instance that can be set to false in order to not check every single time.
See hhucn/adhocracy.hhu_theme#401 for context.