Split user settings in multiple pages (fixes #174)
This splits the user settings page in multiple pages:

* personal settings
* login settings
* notification settings
* advanced settings

This is done the same way as the instance settings are already done.

Most of the forms are refactored towards form element creation using
`templates.form`, some aren't yet as they're more complicated. Some
styling introduced in #456 has been removed again, as it looked bad
within the new layout and should be resolved generically instead.

The gender field adopted minor changes from the ongoing discussion in
#468, as this is in every case better than what was there before.

The user actions have been moved to the *advanced* submenu, but are
still displayed on top, as stupid as before.