Delay asynchronous update (#358)
This allows to delay asynchronous updates by a configurable amount of
time to avoid triggering a race condition between main and background
processes described in #358.

We have used that workaround for quite some time and it seems pragmatic
to codify and enable it by default.

This lowers the priority of this issue from critical to normal.