geo: Update OSM attribution link and style
Attribution now links to www.openstreetmap.org/copyright as recommended.
Previously we explicitly linked to the CC-SA license. As we don't do
this anymore, this is now also usable in installations using data from
ODbL licensed OSM material.

In addition, attribution notes are now displayed nicely within the map.

This is related to #93, which additionally proposes to specify
attribution per map layer.