Allow to freeze single proposals (fixes #559)
This allows to mark proposals as frozen for users with a newley
introduced `proposal.freeze` permission, which is by default assigned to
instance admins.

If a proposal is frozen, it isn't possible

* to edit the proposal (apart from instance admins),
* to comment the proposal (apart from instance admins),
* to rate / to delegate the proposal.

It's also not possible to rate selections, while it is still possible to
edit selections and to comment on selections. These restrictions will be
implemented after merging the sectionpages branch.