fix utility heading margin collapse
When there are actions after a heading the complete thing was
wrappered inside a `<div class="utility">` element and some css
was applied.

The important css was `display: inline-block` so that the
actions were displayed on the same line. This had the side-
effect that margin did not collapse. The head line was followed
by a `<hr>` in most cases the hack `margin-bottom: 0` worked
quite well. However this is not the case with the
show_sectionpage template.

I fixed this by wrapping the actions in a `.utility` *inside*
the heading and adopting the css code. That way the heading css
is not changed and margin collapsing still works.

A similar problem existet with morelink headers and was fixed
in a similar way.

Updates to show_sectionpage template are included in the
following commit