Remove "Event.notifications" relation backref to avoid duplicate notifications
The notification system works in a way that first all possible
notifications are created and later filtered out due to prioritization
etc.

As notifications are now a database mapped class, we need to be careful
to not add them to the SQLAlchemy session by accident.

The mapper backref Event.notifications did precisely that, it
"accidently" added the notifications corresponding to an event to the
database, so this backref has to be removed.