Dashboard: Show comment preview only if it is a comment
We somehow have events corresponding to comments in the database, but
the event data doesn't refer to an actual comment ref, but is
"Undefined" instead. This happens with events related to "issues" (some
former adhocracy object type)...

Anyways it makes sense to be more defensive here, as this would throw
errors on long-serving installations.