Define min and max zoom level per layer
Previously this has been defined for the whole map at once. We now use
the intersection of the allowed zoom levels of a layer and the requested
zoom levels for the whole map.

This also works on image layers, i.e. min and max zoom level can be
configured per image layer in the adhocracy configuration.