Pull in English translations from en_GB in Transifex (fixes #102)
As Transifex doesn't allow to translate the source language, it wasn't
possible to translate English strings through Transifex, which lead to
an outdated English translation. We mostly avoided translating source
strings as this would break the other language's translations.

We originally planned to change the source language to something
artificial (e.g. en_AU), and use en as a proper translatable language.
However this would require deleting and readding the Adhocracy transifex
project entirely, as Transifex doesn't allow changing the source
language. As this would probably mean losing configution and maybe
collaborator, this is a nogo.

This commit fixes this issue in a much more simple way by forcing the
transifex client to also pull the en_GB translations, but store it in
the en folder.

[0] http://support.transifex.com/customer/portal/articles/1170853-can-i
-change-the-source-language-of-my-project-