Use get_string to create connection to solr
If a unicode string is passed to solr, this makes httplib fail at some
point, where message_body (containing a utf-8 encoded string) is added
to msg (which was then a unicode object).

This is fixed by making sure the solr connection url is a string, not a
unicode object.