Workaround: Also pass error message through environment
Before that commit, error messages were passed through the HTTP response
using some webob hack. This however doesn't allow to add arbitrary HTML
(e.g. links).

Therefore this commit passes the error message through the WSGI
environment additionally in order to allow passing arbitrary HTML.

An alternative to that would be to use the webob mechanism and unencode
the HTML in the error controller.

I don't really like either option, but this one is rather simple, so I
prefered that.