Add note about user session loss on domain change
User sessions are tracked through cookies. As cookies are only valid for
the domain of the server sending the cookie, users lose their session
when switching from one instance to another if they have a different
domain.

This could be worked around by requesting some third-party cookies with
javascript. This might be implemented later.